# docs/swagger.yaml
basePath: /v1
definitions:
  api.AddWidgetRequest:
    properties:
      device_id:
        type: string
      position_x:
        type: number
      position_y:
        type: number
      tag_name:
        type: string
      widget_type:
        type: string
    type: object
...
  api.CreateDiagramRequest:
    properties:
      background_url:
        type: string
      name:
        type: string
    type: object
...
  api.UpdateWidgetRequest:
    properties:
      device_id:
        type: string
      position_x:
        type: number
      position_y:
        type: number
      tag_name:
        type: string
      widget_type:
        type: string
    type: object
...
  models.PidDiagram:
    properties:
      background_url:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      site_id:
        type: string
      updated_at:
        type: string
    type: object
  models.PidWidget:
    properties:
      device_id:
        type: string
      diagram_id:
        type: string
      id:
        type: string
      position_x:
        type: number
      position_y:
        type: number
      tag_name:
        type: string
      widget_type:
        description: TEXT, PUMP, VALVE
        type: string
    type: object
...
  response.Meta:
    properties:
      pagination:
        allOf:
        - $ref: '#/definitions/response.Pagination'
        description: Chỉ xuất hiện khi có phân trang, nếu không có sẽ tự động ẩn đi
      timestamp:
        type: string
    type: object
  response.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  response.SuccessResponse:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    name: Linh Community
    url: https://github.com/linhcommunity/nlscada-cloud
  description: Backend API for NL SCADA Cloud
  title: NL SCADA Center API
  version: "1.0"
paths:
...
  /sites/{siteID}/pid-diagrams:
    get:
      parameters:
      - description: Site UUID
        in: path
        name: siteID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.PidDiagram'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Danh sách sơ đồ P&ID
      tags:
      - P&ID
    post:
      consumes:
      - application/json
      parameters:
      - description: Site UUID
        in: path
        name: siteID
        required: true
        type: string
      - description: Thông tin sơ đồ
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.CreateDiagramRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.PidDiagram'
              type: object
      security:
      - BearerAuth: []
      summary: Tạo sơ đồ P&ID
      tags:
      - P&ID
  /sites/{siteID}/pid-diagrams/{diagramID}:
    delete:
      parameters:
      - description: Site UUID
        in: path
        name: siteID
        required: true
        type: string
      - description: Diagram UUID
        in: path
        name: diagramID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Xóa sơ đồ P&ID
      tags:
      - P&ID
    get:
      parameters:
      - description: Site UUID
        in: path
        name: siteID
        required: true
        type: string
      - description: Diagram UUID
        in: path
        name: diagramID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.PidDiagram'
              type: object
      security:
      - BearerAuth: []
      summary: Chi tiết sơ đồ P&ID
      tags:
      - P&ID
  /sites/{siteID}/pid-diagrams/{diagramID}/widgets:
    post:
      consumes:
      - application/json
      parameters:
      - description: Site UUID
        in: path
        name: siteID
        required: true
        type: string
      - description: Diagram UUID
        in: path
        name: diagramID
        required: true
        type: string
      - description: Thông tin widget
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.AddWidgetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.PidWidget'
              type: object
      security:
      - BearerAuth: []
      summary: Thêm widget vào sơ đồ
      tags:
      - P&ID
  /sites/{siteID}/pid-diagrams/{diagramID}/widgets/{widgetID}:
    delete:
      parameters:
      - description: Site UUID
        in: path
        name: siteID
        required: true
        type: string
      - description: Diagram UUID
        in: path
        name: diagramID
        required: true
        type: string
      - description: Widget UUID
        in: path
        name: widgetID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Xóa widget
      tags:
      - P&ID
    put:
      consumes:
      - application/json
      parameters:
      - description: Site UUID
        in: path
        name: siteID
        required: true
        type: string
      - description: Diagram UUID
        in: path
        name: diagramID
        required: true
        type: string
      - description: Widget UUID
        in: path
        name: widgetID
        required: true
        type: string
      - description: Các trường cần cập nhật
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.UpdateWidgetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.PidWidget'
              type: object
      security:
      - BearerAuth: []
      summary: Cập nhật widget
      tags:
      - P&ID
...
securityDefinitions:
  BearerAuth:
    description: Sử dụng "Bearer <token>", nhưng cookie session_token cũng được chấp
      nhận khi gọi từ browser
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"